requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
flask-login==0.6.3
flask-mail==0.9.1
openai==1.61.0
python-dotenv==1.0.1
mistralai==1.5.0
PyPDF2==3.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.23.26
Pillow>=11.0.0
python-dateutil>=2.8.2
pytz>=2024.1
numpy>=1.24.0
pandas>=2.0.0
psycopg2-binary>=2.9.10